docs/coverage/node/html/test/wpt/xhr/send-blob-with-no-mime-type.js.gcov.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">

<html lang="en">

<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <title>LCOV - lcov.info - test/wpt/xhr/send-blob-with-no-mime-type.js</title>
  <link rel="stylesheet" type="text/css" href="../../../gcov.css">
</head>

<body>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="title">LCOV - code coverage report</td></tr>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>

    <tr>
      <td width="100%">
        <table cellpadding=1 border=0 width="100%">
          <tr>
            <td width="10%" class="headerItem">Current view:</td>
            <td width="35%" class="headerValue"><a href="../../../index.html">top level</a> - <a href="index.html">test/wpt/xhr</a> - send-blob-with-no-mime-type.js<span style="font-size: 80%;"> (source / <a href="send-blob-with-no-mime-type.js.func-sort-c.html">functions</a>)</span></td>
            <td width="5%"></td>
            <td width="15%"></td>
            <td width="10%" class="headerCovTableHead">Hit</td>
            <td width="10%" class="headerCovTableHead">Total</td>
            <td width="15%" class="headerCovTableHead">Coverage</td>
          </tr>
          <tr>
            <td class="headerItem">Test:</td>
            <td class="headerValue">lcov.info</td>
            <td></td>
            <td class="headerItem">Lines:</td>
            <td class="headerCovTableEntry">117</td>
            <td class="headerCovTableEntry">117</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td class="headerItem">Date:</td>
            <td class="headerValue">2025-01-24 02:02:48</td>
            <td></td>
            <td class="headerItem">Functions:</td>
            <td class="headerCovTableEntry">5</td>
            <td class="headerCovTableEntry">5</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td></td>
            <td></td>
            <td></td>
            <td class="headerItem">Branches:</td>
            <td class="headerCovTableEntry">7</td>
            <td class="headerCovTableEntry">7</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr><td><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
        </table>
      </td>
    </tr>

    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
  </table>

  <table cellpadding=0 cellspacing=0 border=0>
    <tr>
      <td><br></td>
    </tr>
    <tr>
      <td>
<pre class="sourceHeading">           Branch data     Line data    Source code</pre>
<pre class="source">
<a name="1"><span class="lineNum">       1 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : import assert from 'node:assert/strict';</span></a>
<a name="2"><span class="lineNum">       2 </span>                :<span class="lineCov">          1 : import { XMLHttpRequest } from '../../../lib/whatwg-xhr.js';</span></a>
<a name="3"><span class="lineNum">       3 </span>                :<span class="lineCov">          1 : </span></a>
<a name="4"><span class="lineNum">       4 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : export default (activeURL) =&gt; {</span></a>
<a name="5"><span class="lineNum">       5 </span>                :<span class="lineCov">          1 : </span></a>
<a name="6"><span class="lineNum">       6 </span>                :<span class="lineCov">          1 :         let blobTests = [</span></a>
<a name="7"><span class="lineNum">       7 </span>                :<span class="lineCov">          1 :                 ['no mime type',      new Blob(['data'])],</span></a>
<a name="8"><span class="lineNum">       8 </span>                :<span class="lineCov">          1 :                 ['invalid mime type', new Blob(['data'], {type: 'Invalid \r\n mime \r\n type'})]</span></a>
<a name="9"><span class="lineNum">       9 </span>                :<span class="lineCov">          1 :         ];</span></a>
<a name="10"><span class="lineNum">      10 </span>                :<span class="lineCov">          1 : </span></a>
<a name="11"><span class="lineNum">      11 </span>           [<span class="branchCov" title="Branch 0 was taken 2 times"> + </span>]:<span class="lineCov">          1 :         blobTests.forEach((item) =&gt; {</span></a>
<a name="12"><span class="lineNum">      12 </span>                :<span class="lineCov">          2 :                 doSyncTest(item, 'POST');</span></a>
<a name="13"><span class="lineNum">      13 </span>                :<span class="lineCov">          2 :                 doAsyncTest(item, 'POST');</span></a>
<a name="14"><span class="lineNum">      14 </span>                :<span class="lineCov">          2 : </span></a>
<a name="15"><span class="lineNum">      15 </span>                :<span class="lineCov">          2 :                 doSyncTest(item, 'PUT');</span></a>
<a name="16"><span class="lineNum">      16 </span>                :<span class="lineCov">          2 :                 doAsyncTest(item, 'PUT');</span></a>
<a name="17"><span class="lineNum">      17 </span>                :<span class="lineCov">          1 :         });</span></a>
<a name="18"><span class="lineNum">      18 </span>                :<span class="lineCov">          1 : </span></a>
<a name="19"><span class="lineNum">      19 </span>           [<span class="branchCov" title="Branch 0 was taken 4 times"> + </span>]:<span class="lineCov">          1 :         function doSyncTest(testItem, method){</span></a>
<a name="20"><span class="lineNum">      20 </span>                :<span class="lineCov">          4 : </span></a>
<a name="21"><span class="lineNum">      21 </span>                :<span class="lineCov">          4 :                 let xhr = new XMLHttpRequest();</span></a>
<a name="22"><span class="lineNum">      22 </span>                :<span class="lineCov">          4 : </span></a>
<a name="23"><span class="lineNum">      23 </span>                :<span class="lineCov">          4 :                 xhr.open(method, `${activeURL}/content.py`, false);</span></a>
<a name="24"><span class="lineNum">      24 </span>                :<span class="lineCov">          4 :                 xhr.send(testItem[1]);</span></a>
<a name="25"><span class="lineNum">      25 </span>                :<span class="lineCov">          4 : </span></a>
<a name="26"><span class="lineNum">      26 </span>                :<span class="lineCov">          4 :                 assert.strictEqual(xhr.getResponseHeader('X-Request-Content-Length'), '4');</span></a>
<a name="27"><span class="lineNum">      27 </span>                :<span class="lineCov">          4 :                 assert.strictEqual(xhr.getResponseHeader('X-Request-Content-Type'), 'NO');</span></a>
<a name="28"><span class="lineNum">      28 </span>                :<span class="lineCov">          4 :         }</span></a>
<a name="29"><span class="lineNum">      29 </span>                :<span class="lineCov">          1 : </span></a>
<a name="30"><span class="lineNum">      30 </span>           [<span class="branchCov" title="Branch 0 was taken 4 times"> + </span>]:<span class="lineCov">          1 :         function doAsyncTest(testItem, method) {</span></a>
<a name="31"><span class="lineNum">      31 </span>                :<span class="lineCov">          4 : </span></a>
<a name="32"><span class="lineNum">      32 </span>                :<span class="lineCov">          4 :                 let xhr = new XMLHttpRequest();</span></a>
<a name="33"><span class="lineNum">      33 </span>                :<span class="lineCov">          4 : </span></a>
<a name="34"><span class="lineNum">      34 </span>           [<span class="branchCov" title="Branch 0 was taken 16 times"> + </span>]:<span class="lineCov">          4 :                 xhr.onreadystatechange = () =&gt; {</span></a>
<a name="35"><span class="lineNum">      35 </span>           [<span class="branchCov" title="Branch 0 was taken 4 times"> + </span>]:<span class="lineCov">         16 :                         if(xhr.readyState == 4){</span></a>
<a name="36"><span class="lineNum">      36 </span>                :<span class="lineCov">          4 : </span></a>
<a name="37"><span class="lineNum">      37 </span>                :<span class="lineCov">          4 :                                 assert.strictEqual(xhr.getResponseHeader('X-Request-Content-Length'), '4');</span></a>
<a name="38"><span class="lineNum">      38 </span>                :<span class="lineCov">          4 :                                 assert.strictEqual(xhr.getResponseHeader('X-Request-Content-Type'), 'NO');</span></a>
<a name="39"><span class="lineNum">      39 </span>                :<span class="lineCov">          4 : </span></a>
<a name="40"><span class="lineNum">      40 </span>                :<span class="lineCov">          4 :                                 xhr.onreadystatechange = null;</span></a>
<a name="41"><span class="lineNum">      41 </span>                :<span class="lineCov">          4 :                         }</span></a>
<a name="42"><span class="lineNum">      42 </span>                :<span class="lineCov">          4 :                 };</span></a>
<a name="43"><span class="lineNum">      43 </span>                :<span class="lineCov">          4 : </span></a>
<a name="44"><span class="lineNum">      44 </span>                :<span class="lineCov">          4 :                 xhr.open(method, `${activeURL}/content.py`, true);</span></a>
<a name="45"><span class="lineNum">      45 </span>                :<span class="lineCov">          4 :                 xhr.send(testItem[1]);</span></a>
<a name="46"><span class="lineNum">      46 </span>                :<span class="lineCov">          4 :         }</span></a>
<a name="47"><span class="lineNum">      47 </span>                :<span class="lineCov">          1 : }</span></a>
<a name="48"><span class="lineNum">      48 </span>                :<span class="lineCov">          1 : ﻿</span></a>
<a name="49"><span class="lineNum">      49 </span>                :<span class="lineCov">          1 : /*</span></a>
<a name="50"><span class="lineNum">      50 </span>                :<span class="lineCov">          1 :  * send-blob-with-no-mime-type.html</span></a>
<a name="51"><span class="lineNum">      51 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="52"><span class="lineNum">      52 </span>                :<span class="lineCov">          1 : </span></a>
<a name="53"><span class="lineNum">      53 </span>                :<span class="lineCov">          1 : &lt;!DOCTYPE html&gt;</span></a>
<a name="54"><span class="lineNum">      54 </span>                :<span class="lineCov">          1 : &lt;html&gt;</span></a>
<a name="55"><span class="lineNum">      55 </span>                :<span class="lineCov">          1 : &lt;head&gt;</span></a>
<a name="56"><span class="lineNum">      56 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#the-send()-method&quot; data-tested-assertations=&quot;following::ol[1]/li[4] following::ol[1]/li[4]/dl[1]/dd[2]/p[3]&quot;/&gt;</span></a>
<a name="57"><span class="lineNum">      57 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#the-status-attribute&quot;  data-tested-assertations=&quot;following::ol[1]/li[3]&quot;/&gt;</span></a>
<a name="58"><span class="lineNum">      58 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#the-responsetype-attribute&quot;  data-tested-assertations=&quot;following::ol[1]/li[4]&quot;/&gt;</span></a>
<a name="59"><span class="lineNum">      59 </span>                :<span class="lineCov">          1 :     &lt;link rel=&quot;help&quot; href=&quot;https://xhr.spec.whatwg.org/#the-response-attribute&quot;  data-tested-assertations=&quot;following::a[contains(@href,'#blob-response-entity-body')]/..&quot;/&gt;</span></a>
<a name="60"><span class="lineNum">      60 </span>                :<span class="lineCov">          1 : </span></a>
<a name="61"><span class="lineNum">      61 </span>                :<span class="lineCov">          1 :     &lt;script src=&quot;/resources/testharness.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="62"><span class="lineNum">      62 </span>                :<span class="lineCov">          1 :     &lt;script src=&quot;/resources/testharnessreport.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="63"><span class="lineNum">      63 </span>                :<span class="lineCov">          1 :     &lt;title&gt;XMLHttpRequest: The send() method: Blob data with no mime type&lt;/title&gt;</span></a>
<a name="64"><span class="lineNum">      64 </span>                :<span class="lineCov">          1 : &lt;/head&gt;</span></a>
<a name="65"><span class="lineNum">      65 </span>                :<span class="lineCov">          1 : </span></a>
<a name="66"><span class="lineNum">      66 </span>                :<span class="lineCov">          1 : &lt;body&gt;</span></a>
<a name="67"><span class="lineNum">      67 </span>                :<span class="lineCov">          1 :     &lt;div id=&quot;log&quot;&gt;&lt;/div&gt;</span></a>
<a name="68"><span class="lineNum">      68 </span>                :<span class="lineCov">          1 : </span></a>
<a name="69"><span class="lineNum">      69 </span>                :<span class="lineCov">          1 :     &lt;script type=&quot;text/javascript&quot;&gt;</span></a>
<a name="70"><span class="lineNum">      70 </span>                :<span class="lineCov">          1 :         var blobTests = [</span></a>
<a name="71"><span class="lineNum">      71 </span>                :<span class="lineCov">          1 :             [&quot;no mime type&quot;, new Blob([&quot;data&quot;])],</span></a>
<a name="72"><span class="lineNum">      72 </span>                :<span class="lineCov">          1 :             [&quot;invalid mime type&quot;, new Blob([&quot;data&quot;], {type: &quot;Invalid \r\n mime \r\n type&quot;})]</span></a>
<a name="73"><span class="lineNum">      73 </span>                :<span class="lineCov">          1 :         ];</span></a>
<a name="74"><span class="lineNum">      74 </span>                :<span class="lineCov">          1 : </span></a>
<a name="75"><span class="lineNum">      75 </span>                :<span class="lineCov">          1 :         function doSyncTest(testItem, method) {</span></a>
<a name="76"><span class="lineNum">      76 </span>                :<span class="lineCov">          1 :             test(function() {</span></a>
<a name="77"><span class="lineNum">      77 </span>                :<span class="lineCov">          1 :                 var xhr = new XMLHttpRequest();</span></a>
<a name="78"><span class="lineNum">      78 </span>                :<span class="lineCov">          1 :                 xhr.open(method, &quot;./resources/content.py&quot;, false);</span></a>
<a name="79"><span class="lineNum">      79 </span>                :<span class="lineCov">          1 :                 xhr.send(testItem[1]);</span></a>
<a name="80"><span class="lineNum">      80 </span>                :<span class="lineCov">          1 : </span></a>
<a name="81"><span class="lineNum">      81 </span>                :<span class="lineCov">          1 :                 assert_equals(xhr.getResponseHeader(&quot;X-Request-Content-Length&quot;), &quot;4&quot;);</span></a>
<a name="82"><span class="lineNum">      82 </span>                :<span class="lineCov">          1 :                 assert_equals(xhr.getResponseHeader(&quot;X-Request-Content-Type&quot;), &quot;NO&quot;);</span></a>
<a name="83"><span class="lineNum">      83 </span>                :<span class="lineCov">          1 :             }, &quot;Synchronous blob loading with &quot; + testItem[0] + &quot; [&quot; + method + &quot;]&quot;);</span></a>
<a name="84"><span class="lineNum">      84 </span>                :<span class="lineCov">          1 :         }</span></a>
<a name="85"><span class="lineNum">      85 </span>                :<span class="lineCov">          1 : </span></a>
<a name="86"><span class="lineNum">      86 </span>                :<span class="lineCov">          1 :         function doAsyncTest(testItem, method) {</span></a>
<a name="87"><span class="lineNum">      87 </span>                :<span class="lineCov">          1 :             var atest = async_test(&quot;Asynchronous blob loading with &quot; + testItem[0] + &quot; [&quot; + method + &quot;]&quot;);</span></a>
<a name="88"><span class="lineNum">      88 </span>                :<span class="lineCov">          1 :             atest.step(function() {</span></a>
<a name="89"><span class="lineNum">      89 </span>                :<span class="lineCov">          1 :                 var xhr = new XMLHttpRequest();</span></a>
<a name="90"><span class="lineNum">      90 </span>                :<span class="lineCov">          1 :                 xhr.onreadystatechange = function() {</span></a>
<a name="91"><span class="lineNum">      91 </span>                :<span class="lineCov">          1 :                     if (xhr.readyState == 4) {</span></a>
<a name="92"><span class="lineNum">      92 </span>                :<span class="lineCov">          1 :                         atest.step(function() {</span></a>
<a name="93"><span class="lineNum">      93 </span>                :<span class="lineCov">          1 :                             assert_equals(xhr.getResponseHeader(&quot;X-Request-Content-Length&quot;), &quot;4&quot;);</span></a>
<a name="94"><span class="lineNum">      94 </span>                :<span class="lineCov">          1 :                             assert_equals(xhr.getResponseHeader(&quot;X-Request-Content-Type&quot;), &quot;NO&quot;);</span></a>
<a name="95"><span class="lineNum">      95 </span>                :<span class="lineCov">          1 :                         });</span></a>
<a name="96"><span class="lineNum">      96 </span>                :<span class="lineCov">          1 :                         atest.done();</span></a>
<a name="97"><span class="lineNum">      97 </span>                :<span class="lineCov">          1 :                     }</span></a>
<a name="98"><span class="lineNum">      98 </span>                :<span class="lineCov">          1 :                 }</span></a>
<a name="99"><span class="lineNum">      99 </span>                :<span class="lineCov">          1 :                 xhr.open(method, &quot;./resources/content.py&quot;, true);</span></a>
<a name="100"><span class="lineNum">     100 </span>                :<span class="lineCov">          1 :                 xhr.send(testItem[1]);</span></a>
<a name="101"><span class="lineNum">     101 </span>                :<span class="lineCov">          1 :             });</span></a>
<a name="102"><span class="lineNum">     102 </span>                :<span class="lineCov">          1 :         }</span></a>
<a name="103"><span class="lineNum">     103 </span>                :<span class="lineCov">          1 : </span></a>
<a name="104"><span class="lineNum">     104 </span>                :<span class="lineCov">          1 :         blobTests.forEach(function(item){</span></a>
<a name="105"><span class="lineNum">     105 </span>                :<span class="lineCov">          1 :             doSyncTest(item, &quot;POST&quot;);</span></a>
<a name="106"><span class="lineNum">     106 </span>                :<span class="lineCov">          1 :             doAsyncTest(item, &quot;POST&quot;);</span></a>
<a name="107"><span class="lineNum">     107 </span>                :<span class="lineCov">          1 : </span></a>
<a name="108"><span class="lineNum">     108 </span>                :<span class="lineCov">          1 :             doSyncTest(item, &quot;PUT&quot;);</span></a>
<a name="109"><span class="lineNum">     109 </span>                :<span class="lineCov">          1 :             doAsyncTest(item, &quot;PUT&quot;);</span></a>
<a name="110"><span class="lineNum">     110 </span>                :<span class="lineCov">          1 :         });</span></a>
<a name="111"><span class="lineNum">     111 </span>                :<span class="lineCov">          1 :     &lt;/script&gt;</span></a>
<a name="112"><span class="lineNum">     112 </span>                :<span class="lineCov">          1 : &lt;/body&gt;</span></a>
<a name="113"><span class="lineNum">     113 </span>                :<span class="lineCov">          1 : &lt;/html&gt;</span></a>
<a name="114"><span class="lineNum">     114 </span>                :<span class="lineCov">          1 : </span></a>
<a name="115"><span class="lineNum">     115 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="116"><span class="lineNum">     116 </span>                :<span class="lineCov">          1 :  * send-blob-with-no-mime-type.html</span></a>
<a name="117"><span class="lineNum">     117 </span>                :<span class="lineCov">          1 :  */</span></a>
</pre>
      </td>
    </tr>
  </table>
  <br>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
  </table>
  <br>

</body>
</html>
